// src/app/Models/BusinesObjectModel.ts
// -----------------------------------------------------------------------
// <copyright file="BusinessObjectModel.ts" company="Soloplan GmbH">
//     Copyright (c) Soloplan GmbH. All rights reserved.
// </copyright>
// -----------------------------------------------------------------------

export interface BusinesObjectModel {
    isActive: boolean;
    id: number;
    number: number;
    matchcode: string;
    designation: string;
}